# AzureDemo/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AzureDemo, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Authentication": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/Authentication/GetUserId": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Authentication/GetUser": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/notifications/register": {
      "get": {
        "tags": [
          "PushNotifications"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/notifications/unregister/{registrationId}": {
      "delete": {
        "tags": [
          "PushNotifications"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "registrationId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/notifications/enable/{id}": {
      "put": {
        "tags": [
          "PushNotifications"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DeviceRegistration"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/notifications/send": {
      "post": {
        "tags": [
          "PushNotifications"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Notification"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/notifications/getPushNotificationByUser": {
      "get": {
        "tags": [
          "PushNotifications"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/notifications/getAllPushNotification": {
      "get": {
        "tags": [
          "PushNotifications"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "DeviceRegistration": {
      "type": "object",
      "properties": {
        "platform": {
          "format": "int32",
          "type": "integer"
        },
        "handle": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Notification": {
      "type": "object",
      "properties": {
        "platform": {
          "format": "int32",
          "type": "integer"
        },
        "handle": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "content": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "emailId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    }
  }
}